+++ azure-iot-device/azure/iot/device/iothub/sync_clients.py
@@ -197,6 +197,37 @@ class GenericIoTHubClient(AbstractIoTHubClient):
 
         logger.info("Successfully disconnected from Hub")
 
+    def update_sastoken(self, sastoken):
+        """
+        Update the client's SAS Token used for authentication, then reauthorizes the connection.
+
+        This API can only be used if the client was initially created with a SAS Token.
+        Note also that this API may return before the reauthorization/reconnection is completed.
+        This means that some errors that may occur as part of the reconnection could occur in the
+        background, and will not be raised by this method.
+
+        :param str sastoken: The new SAS Token string for the client to use
+
+        :raises: :class:`azure.iot.device.exceptions.ClientError` if the client was not initially
+            created with a SAS token.
+        :raises: :class:`azure.iot.device.exceptions.ClientError` if there is an unexpected failure
+            during execution.
+        :raises: ValueError if the sastoken parameter is invalid
+        """
+        self._replace_user_supplied_sastoken(sastoken)
+
+        # Reauthorize the connection
+        logger.info("Reauthorizing connection with Hub...")
+        callback = EventedCallback()
+        self._mqtt_pipeline.reauthorize_connection(callback=callback)
+        handle_result(callback)
+        # NOTE: Currently due to the MQTT3 implemenation, the pipeline reauthorization will return
+        # after the disconnect. It does not wait for the reconnect to complete. This means that
+        # any errors that may occur as part of the connect will not return via this callback.
+        # They will instead go to the background exception handler.
+
+        logger.info("Successfully reauthorized connection to Hub")
+
     def send_message(self, message):
         """Sends a message to the default events endpoint on the Azure IoT Hub or Azure IoT Edge Hub instance.
