+++ credscan_suppression.json
@@ -22,7 +22,11 @@
       "_justification": "Test containing fake keys and fake signed data"
     },
     {
-      "file": "\\azure-iot-device\\tests\\iothub\\shared_client_tests.py",
+      "file": "\\azure-iot-device\\tests\\iothub\\test_sync_clients.py",
+      "_justification": "Test containing fake signed data"
+    },
+    {
+      "file": "\\azure-iot-device\\tests\\iothub\\aio\\test_async_clients.py",
       "_justification": "Test containing fake signed data"
     }
   ]
